test(zai): isolate stream boundary fixture
